refactor(runtime): extract transport coordination module (#3059)
